+++ lib/models/index.dart
@@ -16,6 +16,7 @@ export 'recipes/recipe.dart';
 export 'recipes/review.dart';
 export 'recipes/time.dart';
 export 'recipes/utensil.dart';
-export 'search/enums.dart';
+export 'search/enums/category.dart';
+export 'search/enums/sort.dart';
 export 'search/result.dart';
 export 'videos/video.dart';
